removing float64 type alias in dtypes since it is redundant

+++ regression-tests/sparktkregtests/testcases/frames/extreme_value_test.py
@@ -31,7 +31,7 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
         self.schema_proj = [("col_A", dtypes.int32),
                             ("col_B", dtypes.int64),
                             ("col_C", dtypes.float32),
-                            ("Double", dtypes.float64),
+                            ("Double", dtypes.float32),
                             ("Text", str)]
 
     def test_extreme_projection(self):
@@ -47,7 +47,7 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
                        4321: np.nan}
             return new_val.get(row["col_A"], row["col_A"])
 
-        master.add_columns(add_extremes, ("col_D", dtypes.float64))
+        master.add_columns(add_extremes, ("col_D", dtypes.float32))
         
         proj_3col = master.copy(['col_D', 'Double', 'Text'])
         self.assertEqual(proj_3col.count(), master.count())
@@ -76,7 +76,7 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
 
         frame = self.context.frame.import_csv(
             self.data_proj, schema=self.schema_proj)
-        frame.add_columns(add_extremes, ("col_D", dtypes.float64))
+        frame.add_columns(add_extremes, ("col_D", dtypes.float32))
 
         frame_copy = frame.copy()
         self.assertEqual(frame.column_names, frame_copy.column_names)
@@ -99,16 +99,16 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
     def test_extreme_maxmin64(self):
         """ Test extreme large and small magnitudes on 64-bit floats."""
         data_maxmin64 = self.get_file('BigAndTinyFloat64s.csv')
-        schema_maxmin64 = [("col_A", dtypes.float64),
-                           ("col_B", dtypes.float64)]
+        schema_maxmin64 = [("col_A", dtypes.float32),
+                           ("col_B", dtypes.float32)]
         extreme64 = self.context.frame.import_csv(
             data_maxmin64, schema=schema_maxmin64)
         self.assertEqual(extreme64.count(), 16)
 
         extreme64.add_columns(lambda row:
                               [row.col_A*2, np.sqrt(-9)],
-                              [("twice", dtypes.float64),
-                               ('neg_root', dtypes.float64)])
+                              [("twice", dtypes.float32),
+                               ('neg_root', dtypes.float32)])
         extake = extreme64.to_pandas(extreme64.count())
 
         #check for inf when values exceed 64-bit range;
@@ -129,10 +129,10 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
             return new_val.get(row["item"], row['weight'])
 
         data_stat = self.get_file("mode_stats.tsv")
-        schema_stat = [("weight", dtypes.float64), ("item", int)]
+        schema_stat = [("weight", dtypes.float32), ("item", int)]
         stat_frame = self.context.frame.import_csv(
             data_stat, schema=schema_stat, delimiter="\t")
-        stat_frame.add_columns(add_extremes, ("weight2", dtypes.float64))
+        stat_frame.add_columns(add_extremes, ("weight2", dtypes.float32))
         stats = stat_frame.column_mode(
             "item", "weight2", max_modes_returned=50)
 
@@ -155,7 +155,7 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
             data_mode, schema=schema_mode)
 
         # Create new column where only values 10 and 50 are valid;
-        stat_frame.add_columns(add_extremes, ("Item2", dtypes.float64))
+        stat_frame.add_columns(add_extremes, ("Item2", dtypes.float32))
         stat_frame.drop_columns("Item")
         stat_frame.rename_columns({"Item2": "Item"})
 
@@ -171,21 +171,21 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
     def test_extreme_col_summary_corner(self):
         """ Test column_summary_stats with very large values and odd cases."""
         def add_extremes(row):
-            bf = dtypes.float64(2) ** 1021
+            bf = dtypes.float32(2) ** 1021
             new_val = {600: 30,
                        1: 0,
                        2: bf * 2,
-                       3: dtypes.float64(2) ** 1023,
+                       3: dtypes.float32(2) ** 1023,
                        3.1: 5.4}
             return new_val.get(row.Factors, row.Factors)
         expected_stats = [1624.6, 56, 563700.64]
         data_corner = self.get_file("SummaryStats2.csv")
-        schema_corner = [("Item", dtypes.float64), ("Factors", dtypes.float64)]
+        schema_corner = [("Item", dtypes.float32), ("Factors", dtypes.float32)]
 
         stat_frame = self.context.frame.import_csv(
             data_corner, schema=schema_corner)
 
-        stat_frame.add_columns(add_extremes, ("Weight", dtypes.float64))
+        stat_frame.add_columns(add_extremes, ("Weight", dtypes.float32))
         stat_frame.drop_columns("Factors")
         stats = stat_frame.column_summary_statistics(
                            "Item", weights_column="Weight")
@@ -195,7 +195,7 @@ class ExtremeValueTest(sparktk_test.SparkTKTestCase):
 
     def test_extreme_col_summary_empty_frame(self):
         expected_stats = [float('nan'), 0, 1.0]
-        schema = [("Item", dtypes.float64), ("Weight", dtypes.float64)]
+        schema = [("Item", dtypes.float32), ("Weight", dtypes.float32)]
         stat_frame = self.context.frame.create([], schema=schema)
         stats = stat_frame.column_summary_statistics(
                            "Item", weights_column="Weight")
